requirements.txt
# Core Testing Framework
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-html>=4.1.0
pytest-xdist>=3.3.1
pytest-timeout>=2.2.0

# UI Testing - Playwright
playwright>=1.40.0
pytest-playwright>=0.4.0

# API Testing
requests>=2.31.0
urllib3>=2.0.0

# Database
pyodbc>=5.0.0
pymysql>=1.1.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
faker>=20.0.0
pyyaml>=6.0.0

# Reporting
allure-pytest>=2.13.0

# Data Handling
pandas>=2.0.0
openpyxl>=3.1.0

# Logging
colorlog>=6.7.0
